Refactor acceptance tests to use separate image for certs
Before this change, there was some tomfoolery around moving the certs onto the
puppetserver, which result in strange and inconsistent behavior when the
certificate file locations were modified while the puppetserver was running.

This change separates out those certs to a specific image that both the vault
and puppetserver images use, making the initialization of puppetserver cleaner
without having to restart it.