Port to use the new http client where available
Puppet added a new http client in version 6.11. There wasn't a
compelling reason to port this, until 7.16 when that http client gained
the ability to use the system CA store. This meant that it could begin
trusting certificates not issued by Puppet's CA.

This conditionally ports to the new client where available, and uses the
system CA store as of 7.16.

Co-authored-by: vamsinm <vamsi.nainavarapu.vaap9t@statefarm.com>
Co-authored-by: Christos Papageorgiou <christos.papageorgioy@gmail.com>
Co-authored-by: Josh Cooper <737664+joshcooper@users.noreply.github.com>