Cache lookup results based on a combination of path, vault_addr, and namespace
This caches the result of a Vault lookup based on the `path`,
`vault_addr`, and `namespace` used for the lookup. All subsequent
lookups of the same secret will be returned from Puppet's built in cache
rather than being fetched from vault again.

The caching implementation uses the cache_param feature of the 4.x
Function API.

Note that the key used to cache the data is built as the following
array:
```
[path, vault_addr, namespace]
```

Using just the `path` as a cache key has the possibility of
inadvertantly returning the wrong value as all three of `path`,
`vault_addr`, and `namespace` could affect which secret is looked up
from Vault.

For example, looking up the `foo` secret in Vault server 1 might return
a different value than looking up `foo` in Vault server 2.
And looking up the same path in different namespaces could return a
different value.